Add elastic search functions into a service
